refactor(vue): align A2UI source hierarchy
Move Vue A2UI integration into the React-equivalent packages/vue/src/v2/a2ui/
layout with vue-renderer/, operations, and types modules while preserving
existing component and package export paths through re-export shims.
