+++ packages/vue/src/v2/a2ui/index.ts
@@ -1,0 +1,23 @@
+export { createA2UIMessageRenderer } from "./A2UIMessageRenderer";
+export type { A2UIMessageRendererOptions } from "./A2UIMessageRenderer";
+export { registerA2UICatalogContext } from "./A2UICatalogContext";
+export { registerA2UIBuiltInToolCallRenderer } from "./A2UIToolCallRenderer";
+export { default as A2UISurfaceActivityRenderer } from "./A2UISurfaceActivityRenderer.vue";
+export {
+  A2UISurfaceActivityType,
+  A2UIActivityContentSchema,
+  getOperationSurfaceId,
+} from "./operations";
+export type { A2UIOperation } from "./operations";
+export type { A2UITheme, A2UISurfaceOperationPayload } from "./types";
+export {
+  createVueComponent,
+  createBinderlessVueComponent,
+  A2uiSurface,
+  DeferredChild,
+  vueBasicCatalog,
+} from "./vue-renderer";
+export type {
+  VueComponentImplementation,
+  VueA2uiComponentProps,
+} from "./vue-renderer";
