+++ packages/vue/src/v2/a2ui/types.ts
@@ -1,0 +1,5 @@
+export type A2UITheme = Record<string, unknown>;
+
+export interface A2UISurfaceOperationPayload {
+  operations: unknown[];
+}
