+++ packages/vue/src/v2/a2ui/vue-renderer/A2uiSurface.ts
@@ -5,20 +5,11 @@
  * @a2ui/web_core's framework-agnostic primitives directly.
  */
 
-import {
-  defineComponent,
-  h,
-  ref,
-  onUnmounted,
-  type PropType,
-  type VNode,
-} from "vue";
-import {
-  ComponentContext,
-  type SurfaceModel,
-  type ComponentModel,
-} from "@a2ui/web_core/v0_9";
-import type { VueComponentImplementation } from "./adapter";
+import { defineComponent, h, ref, onUnmounted } from "vue";
+import type { PropType, VNode } from "vue";
+import { ComponentContext } from "@a2ui/web_core/v0_9";
+import type { SurfaceModel, ComponentModel } from "@a2ui/web_core/v0_9";
+import type { VueComponentImplementation } from "../vue-renderer/adapter";
 
 /**
  * DeferredChild — Vue equivalent of the React DeferredChild.
