+++ packages/vue/src/v2/a2ui/vue-renderer/index.ts
@@ -1,0 +1,7 @@
+export { createVueComponent, createBinderlessVueComponent } from "./adapter";
+export type {
+  VueComponentImplementation,
+  VueA2uiComponentProps,
+} from "./adapter";
+export { A2uiSurface, DeferredChild } from "./A2uiSurface";
+export { vueBasicCatalog } from "./catalog/basic";
