+++ packages/vue/src/v2/components/a2ui/index.ts
@@ -1,7 +1,1 @@
-export { createVueComponent, createBinderlessVueComponent } from "./adapter";
-export type {
-  VueComponentImplementation,
-  VueA2uiComponentProps,
-} from "./adapter";
-export { vueBasicCatalog } from "./catalog";
-export { A2uiSurface, DeferredChild } from "./VueSurface";
+export * from "../../a2ui/vue-renderer";
