+++ packages/vue/src/v2/components/index.ts
@@ -1,7 +1,10 @@
 export { default as CopilotKitInspector } from "./CopilotKitInspector.vue";
 export * from "./MCPAppsActivityRenderer";
-export { default as A2UISurfaceActivityRenderer } from "./A2UISurfaceActivityRenderer.vue";
-export * from "./A2UIMessageRenderer";
+export {
+  A2UISurfaceActivityRenderer,
+  createA2UIMessageRenderer,
+} from "../a2ui";
+export type { A2UIMessageRendererOptions } from "../a2ui";
 export * from "./OpenGenerativeUIRenderer";
 export * from "./a2ui";
 export * from "./chat";
