+++ packages/vue/src/v2/index.ts
@@ -8,7 +8,7 @@ export * from "./components";
 // named export. Vue users need a catalog to pass to `a2ui.catalog` for the
 // catalog-on-provider path; the nested `export *` barrel above gets
 // tree-shaken by the library build, so surface it directly here.
-export { vueBasicCatalog } from "./components/a2ui/catalog";
+export { vueBasicCatalog } from "./a2ui/vue-renderer/catalog/basic";
 export * from "./hooks";
 export * from "./providers";
 export * from "./types";
