+++ packages/vue/src/v2/types/a2ui.ts
@@ -1,5 +1,1 @@
-export type A2UITheme = Record<string, unknown>;
-
-export interface A2UISurfaceOperationPayload {
-  operations: unknown[];
-}
+export * from "../a2ui/types";
